+++ Rakefile
@@ -134,9 +134,9 @@ namespace :docs do
       Primer::Button::Base,
       Primer::Button::Close,
       Primer::Button::Link,
+      Primer::Button::Marketing,
       Primer::ButtonComponent,
       Primer::ButtonGroupComponent,
-      Primer::ButtonMarketingComponent,
       Primer::CounterComponent,
       Primer::DetailsComponent,
       Primer::DropdownComponent,
