+++ docs/content/components/buttonmarketing.md
@@ -1,7 +1,7 @@
 ---
 title: ButtonMarketing
 status: Alpha
-source: https://github.com/primer/view_components/tree/main/app/components/primer/button_marketing_component.rb
+source: https://github.com/primer/view_components/tree/main/app/components/primer/marketing.rb
 storybook: https://primer.style/view-components/stories/?path=/story/primer-button-marketing-component
 ---
 
@@ -18,11 +18,11 @@ Use buttons for actions (e.g. in forms). Use links for destinations, or moving f
 <Example src="<button type='button' class='btn-mktg mr-2'>Default</button><button type='button' class='btn-mktg btn-primary-mktg mr-2'>Primary</button><button type='button' class='btn-mktg btn-outline-mktg '>Outline</button><div class='color-bg-canvas-inverse'>  <button type='button' class='btn-mktg btn-transparent '>Transparent</button></div>" />
 
 ```erb
-<%= render(Primer::ButtonMarketingComponent.new(mr: 2)) { "Default" } %>
-<%= render(Primer::ButtonMarketingComponent.new(scheme: :primary, mr: 2)) { "Primary" } %>
-<%= render(Primer::ButtonMarketingComponent.new(scheme: :outline)) { "Outline" } %>
+<%= render(Primer::Button::Marketing.new(mr: 2)) { "Default" } %>
+<%= render(Primer::Button::Marketing.new(scheme: :primary, mr: 2)) { "Primary" } %>
+<%= render(Primer::Button::Marketing.new(scheme: :outline)) { "Outline" } %>
 <div class="color-bg-canvas-inverse">
-  <%= render(Primer::ButtonMarketingComponent.new(scheme: :transparent)) { "Transparent" } %>
+  <%= render(Primer::Button::Marketing.new(scheme: :transparent)) { "Transparent" } %>
 </div>
 ```
 
@@ -31,8 +31,8 @@ Use buttons for actions (e.g. in forms). Use links for destinations, or moving f
 <Example src="<button type='button' class='btn-mktg mr-2'>Default</button><button type='button' class='btn-mktg btn-large-mktg '>Large</button>" />
 
 ```erb
-<%= render(Primer::ButtonMarketingComponent.new(mr: 2)) { "Default" } %>
-<%= render(Primer::ButtonMarketingComponent.new(variant: :large)) { "Large" } %>
+<%= render(Primer::Button::Marketing.new(mr: 2)) { "Default" } %>
+<%= render(Primer::Button::Marketing.new(variant: :large)) { "Large" } %>
 ```
 
 ## Arguments
