+++ static/arguments.yml
@@ -218,6 +218,29 @@
     type: Boolean
     default: N/A
     description: 'Whether button is full-width with `display: block`.'
+- component: ButtonMarketing
+  source: https://github.com/primer/view_components/tree/main/app/components/primer/marketing.rb
+  parameters:
+  - name: scheme
+    type: Symbol
+    default: "`:default`"
+    description: One of `:default`, `:primary`, `:outline`, or `:transparent`.
+  - name: variant
+    type: Symbol
+    default: "`:default`"
+    description: One of `:default` and `:large`.
+  - name: tag
+    type: Symbol
+    default: "`:button`"
+    description: One of `:button` and `:a`.
+  - name: type
+    type: Symbol
+    default: "`:button`"
+    description: One of `:button` and `:submit`.
+  - name: system_arguments
+    type: Hash
+    default: N/A
+    description: "[System arguments](/system-arguments)"
 - component: Button
   source: https://github.com/primer/view_components/tree/main/app/components/primer/button_component.rb
   parameters:
@@ -248,29 +271,6 @@
 - component: ButtonGroup
   source: https://github.com/primer/view_components/tree/main/app/components/primer/button_group_component.rb
   parameters:
-  - name: system_arguments
-    type: Hash
-    default: N/A
-    description: "[System arguments](/system-arguments)"
-- component: ButtonMarketing
-  source: https://github.com/primer/view_components/tree/main/app/components/primer/button_marketing_component.rb
-  parameters:
-  - name: scheme
-    type: Symbol
-    default: "`:default`"
-    description: One of `:default`, `:primary`, `:outline`, or `:transparent`.
-  - name: variant
-    type: Symbol
-    default: "`:default`"
-    description: One of `:default` and `:large`.
-  - name: tag
-    type: Symbol
-    default: "`:button`"
-    description: One of `:button` and `:a`.
-  - name: type
-    type: Symbol
-    default: "`:button`"
-    description: One of `:button` and `:submit`.
   - name: system_arguments
     type: Hash
     default: N/A
