--- stories/primer/button_marketing_component_stories.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-class Primer::ButtonMarketingComponentStories < ViewComponent::Storybook::Stories
-  layout "storybook_preview"
-
-  story(:button_marketing) do
-    controls do
-      select(:scheme, Primer::ButtonMarketingComponent::SCHEME_OPTIONS, :primary)
-      select(:variant, Primer::ButtonMarketingComponent::VARIANT_OPTIONS, :default)
-      select(:tag, Primer::ButtonMarketingComponent::TAG_OPTIONS, :button)
-      select(:type, Primer::ButtonMarketingComponent::TYPE_OPTIONS, :button)
-    end
-
-    content do
-      "Click me"
-    end
-  end
-end
